style: rename src/config/env -> src/config/config
