+++ src/features/home/api/getScenicSpotAPI.js
@@ -1,7 +1,7 @@
 import { i18n } from 'next-i18next';
 import axios from 'axios';
 
-import { API_HOSTNAME_URL } from 'src/config/env';
+import { API_HOSTNAME_URL } from 'src/config/config';
 import { getAuthorizationHeader } from '@utils/getAuthorizationHeader';
 
 /** ---------------------------------------------------------------------------------------------------------------------
