+++ src/utils/seo.js
@@ -1,7 +1,7 @@
 import Head from 'next/head';
 import { useRouter } from 'next/router';
 
-import { WEB_HOSTNAME_URL } from 'src/config/env';
+import { WEB_HOSTNAME_URL } from 'src/config/config';
 import siteMetaData from 'src/config/siteMetaData';
 
 /**
